+++ examples/scenes/src/Scenes.java
@@ -6,7 +6,7 @@ import org.jetbrains.skija.*;
 
 public class Scenes {
     public static TreeMap<String, Scene> scenes;
-    public static String currentScene = "Bitmap";
+    public static String currentScene = "Shadows";
     public static HUD hud = new HUD();
     public static boolean stats = true;
 
@@ -46,6 +46,7 @@ public class Scenes {
         scenes.put("SVG", null);
         scenes.put("SVG Scaling", null);
         scenes.put("Shaders", null);
+        scenes.put("Shadows", null);
         scenes.put("Shadow Utils", null);
         scenes.put("Shapers", null);
         scenes.put("Skottie", null);